requirements.txt
torch
torchvision
sentence-transformers[image]
faiss-cpu
pillow
numpy
tqdm
